requirements/local.txt
asgiref
Django==4.1
django-model-utils
djangorestframework==3.14
Pillow
psycopg2-binary
pytz
sqlparse
Unipath
